dalp-generate-pdf/requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
azure-core==1.3.0
azure-functions==1.2.0
azure-storage-blob==12.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cryptography==2.8
fpdf==1.7.2
idna==2.9
isodate==0.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
msrest==0.6.11
oauthlib==3.1.0
pathspec==0.7.0
pycodestyle==2.5.0
pycparser==2.20
pylint==2.4.4
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
